Add Initial Lookup Functionality
Actions:

+ Added files that contain hashes of encoded polygon
data for the three types of county.

+ Added uk_county_locator wrapped module that kicks
off logic.

+ Added a locator class that encapsulates the logic
to find a polygon from passed in coordinates.

+ Added geokit, parallel and polylines gem dependencies.

+ Added polylines_service class that encapsulates
all of the logic to decode polygon strings and
determine whether a point falls within them.

+ Added specs, fixtures to test each polygon and
a shared example to iterate through a fixture for
a county type.

+ Set logic, so that .find_county will return a
ceremonial county by default, but the user can
pass in a type to return either historic or
current, or all three in a hash.

Motivation:

+ Found that Geokit gem was the fastest to decode
the polylines.

+ Using parallel to work through the polygons in
parallel and return a result faster.

+ Quicker to return a single county, and, after
consideration, thought it most likely that a user
would want to fetch a Ceremonial County.